# Conflicts:
#	tests/integration/cases/dict_with_control_char_keys/Occam@v2.occ
#	tests/integration/cases/string_control_chars/Occam@v2.occ

+++ tests/integration/cases/set_mixed_int_widths/Go_type_hints_safe@v1_18.go
@@ -1,7 +1,7 @@
 package main
 
 func main() {
-my_data := map[int]struct{}{
+my_data := map[int64]struct{}{
 	1: struct{}{},
 	1099511627776: struct{}{},
 }
